Port over the stub controllers implementation. Again, don't pass the full 32-bit address. The component should have no knowledge of the surrounding address space, except that which is strictly necessary. Just pass in a byte therefore.

+++ philpsx_core/src/spu.rs
@@ -10,8 +10,8 @@ pub mod psx_spu;
 pub trait Spu {
 
     /// This must be called in order to read a byte from the SPU.
-    fn read_byte(&self, address: usize) -> i8;
+    fn read_byte(&self, address: i32) -> i8;
 
     /// This must be called in order to write a byte to the SPU.
-    fn write_byte(&mut self, address: usize, value: i8);
+    fn write_byte(&mut self, address: i32, value: i8);
 }
